+++ Cargo.toml
@@ -1,26 +1,9 @@
 [package]
 name = "compile_reproducer"
 version = "0.1.0"
-edition = "2021"
+edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-cortex-m = "0.7.3"
-cortex-m-rt = "0.7.0"
-
 [profile.dev]
-codegen-units = 1
-debug = 2
-debug-assertions = true
-incremental = false
 opt-level = 3
-overflow-checks = true
-
-# do not optimize proc-macro crates = faster builds from scratch
-[profile.dev.build-override]
-codegen-units = 8
-debug = false
-debug-assertions = false
-opt-level = 0
-overflow-checks = false
