+++ Readme.md
@@ -3,5 +3,5 @@
 
 ```
 rustup target add thumbv6m-none-eabi
-cargo run
+cargo build
 ```
